+++ src/clj/rems/api/user_settings.clj
@@ -3,7 +3,15 @@
             [rems.api.schema :as schema]
             [rems.db.user-settings :as user-settings]
             [rems.util :refer [getx-user-id get-user-id]]
-            [ring.util.http-response :refer :all]))
+            [ring.util.http-response :refer :all]
+            [schema.core :as s])
+  (:import (org.joda.time DateTime)))
+
+(def GetUserSettings user-settings/UserSettings)
+
+(s/defschema UpdateUserSettings
+  {(s/optional-key :language) s/Keyword
+   (s/optional-key :notification-email) (s/maybe s/Str)})
 
 (def user-settings-api
   (context "/user-settings" []
@@ -12,12 +20,12 @@
     (GET "/" []
       :summary "Get user settings"
       :roles #{:logged-in}
-      :return user-settings/UserSettings
+      :return GetUserSettings
       (ok (user-settings/get-user-settings (get-user-id))))
 
     (PUT "/" []
       :summary "Update user settings"
       :roles #{:logged-in}
-      :body [settings user-settings/PartialUserSettings]
+      :body [settings UpdateUserSettings]
       :return schema/SuccessResponse
       (ok (user-settings/update-user-settings! (getx-user-id) settings)))))