+++ test/clj/rems/db/test_user_settings.clj
@@ -1,12 +1,16 @@
 (ns ^:integration rems.db.test-user-settings
   (:require [clojure.test :refer :all]
+            [clj-time.core :as time-core]
             [rems.db.testing :refer [rollback-db-fixture test-db-fixture]]
+            [rems.config :as config]
             [rems.db.user-settings :as user-settings]
             [rems.db.users :as users]
             [schema.core :as s]))
 
 (use-fixtures :once test-db-fixture)
 (use-fixtures :each rollback-db-fixture)
+
+(def expiration-date (time-core/date-time 2021 3 18))
 
 (deftest test-user-settings
   (users/add-user! {:userid "user"})
@@ -18,25 +22,39 @@
            (s/validate user-settings/UserSettings
                        (user-settings/get-user-settings "user")))))
 
+  (testing "add settings without EGA flag"
+    (is (= {:success false}
+           (with-redefs [config/env {:enable-ega false}]
+             (user-settings/update-user-settings! "user" {:language :fi
+                                                          :ega {:api-key-expiration-date expiration-date}}))))
+    (is (= {:language :en
+            :notification-email nil}
+           (user-settings/get-user-settings "user"))
+        "nothing was saved since something was wrong"))
+
   (testing "add settings"
     (is (= {:success true}
-           (user-settings/update-user-settings! "user" {:language :fi})))
+           (user-settings/update-user-settings! "user" {:language :fi
+                                                        :ega {:api-key-expiration-date expiration-date}})))
     (is (= {:language :fi
-            :notification-email nil}
+            :notification-email nil
+            :ega {:api-key-expiration-date expiration-date}}
            (user-settings/get-user-settings "user"))))
 
   (testing "modify settings"
     (is (= {:success true}
            (user-settings/update-user-settings! "user" {:notification-email "user@example.com"})))
     (is (= {:language :fi
-            :notification-email "user@example.com"}
+            :notification-email "user@example.com"
+            :ega {:api-key-expiration-date expiration-date}}
            (user-settings/get-user-settings "user"))))
 
   (testing "updating with empty settings does not change settings"
     (is (= {:success true}
            (user-settings/update-user-settings! "user" {})))
     (is (= {:language :fi
-            :notification-email "user@example.com"}
+            :notification-email "user@example.com"
+            :ega {:api-key-expiration-date expiration-date}}
            (user-settings/get-user-settings "user"))))
 
   (testing "updating with invalid settings does not change settings"
@@ -48,7 +66,8 @@
                                                         :language :de})) ; invalid
         "partially invalid")
     (is (= {:language :fi
-            :notification-email "user@example.com"}
+            :notification-email "user@example.com"
+            :ega {:api-key-expiration-date expiration-date}}
            (user-settings/get-user-settings "user"))))
 
   (testing "changing one user's settings does not change another user's settings"
@@ -66,10 +85,10 @@
     (is (= {:language :fi} (user-settings/validate-new-settings {:language :fi}))))
 
   (testing "undefined language"
-    (is (= {} (user-settings/validate-new-settings {:language :de}))))
+    (is (= nil (user-settings/validate-new-settings {:language :de}))))
 
   (testing "nil language"
-    (is (= {} (user-settings/validate-new-settings {:language nil})))))
+    (is (= nil (user-settings/validate-new-settings {:language nil})))))
 
 (deftest test-validate-new-settings
   (testing "valid email"
@@ -77,7 +96,7 @@
            (user-settings/validate-new-settings {:notification-email "somebody@example.com"}))))
 
   (testing "invalid email"
-    (is (= {}
+    (is (= nil
            (user-settings/validate-new-settings {:notification-email "somebody"}))))
 
   (testing "empty email will clear the setting"